+++ test/widget/main_shell_test.dart
@@ -5,6 +5,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:legado_flutter/database/dao/book_dao.dart';
 import 'package:legado_flutter/database/dao/replace_rule_dao.dart';
 import 'package:legado_flutter/database/dao/source_dao.dart';
+import 'package:legado_flutter/application/startup/startup_task_runner.dart';
 import 'package:legado_flutter/domain/crash/crash_report.dart';
 import 'package:legado_flutter/domain/ports/public_text_fetch_port.dart';
 import 'package:legado_flutter/infrastructure/cache/file_chapter_content_cache.dart';
@@ -70,6 +71,7 @@ void main() {
           Provider<PublicTextFetchPort>.value(
             value: const _EmptyPublicTextFetchPort(),
           ),
+          Provider<StartupTaskRunner>(create: (_) => StartupTaskRunner()),
           Provider<BookSourceService>(
             create: (_) => createTestBookSourceService(),
           ),
@@ -130,6 +132,7 @@ void main() {
           Provider<PublicTextFetchPort>.value(
             value: const _EmptyPublicTextFetchPort(),
           ),
+          Provider<StartupTaskRunner>(create: (_) => StartupTaskRunner()),
           Provider<BookSourceService>(
             create: (_) => createTestBookSourceService(),
           ),
@@ -189,6 +192,7 @@ void main() {
             Provider<PublicTextFetchPort>.value(
               value: const _EmptyPublicTextFetchPort(),
             ),
+            Provider<StartupTaskRunner>(create: (_) => StartupTaskRunner()),
             Provider<BookSourceService>(
               create: (_) => createTestBookSourceService(),
             ),